migrate to libxmp
App was using GStreamer through Phonon in past. But modplug plugin was
not always provided by distributions so it had to go away.

Michał Schulz suggested move to libxmp so I did. More supported formats
as a bonus ;D

UI hangs when music is played. ALSA is used directly instead of Phonon
or QtMultimedia.